+++ tasks/write-stage-badges.mjs
@@ -37,5 +37,3 @@ await Promise.all(features.map(async feature => {
 		fs.writeFile(badgeURLOld, shield)
 	]);
 }));
-
-process.exit(0);